requirements/documentation.txt
# Documentation
# -------------

myst-parser[linkify]>=0.14,<0.19
pip-licenses>=3.5,<3.6
sphinx-autobuild==2021.3.14
sphinx-autodoc-typehints<1.19
sphinx-book-theme<1
sphinx-click<4.3
sphinx-copybutton<1
sphinxext-opengraph>=0.4,<1
